\name{proyecto_master-package}
\alias{proyecto_master-package}
\alias{proyecto_master}
\docType{package}
\title{
\packageTitle{proyecto_master}
}
\description{
\packageDescription{proyecto_master}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{proyecto_master}
\packageIndices{proyecto_master}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{proyecto_master}

Maintainer: \packageMaintainer{proyecto_master}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
